pyproject.toml
[project]
name = "frontend-tester"
version = "0.1.0"
description = "AI-powered CLI tool for automated blackbox regression testing of frontend applications"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Frontend Tester Contributors" }
]
keywords = ["testing", "frontend", "regression", "playwright", "bdd", "gherkin", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI Framework
    "typer>=0.12.0",
    "rich>=13.7.0",

    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",

    # Templates
    "jinja2>=3.1.2",

    # BDD Framework
    "pytest-bdd>=7.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",  # Required for async fixtures and step definitions
    "assertpy>=1.1",  # Fluent assertions for tests

    # Browser Automation (Phase 2)
    "playwright>=1.40.0",

    # AI Test Generation (Phase 3)
    "litellm>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "assertpy>=1.1",
]

ai = [
    "litellm>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]

visual = [
    "pillow>=10.0.0",
    "pixelmatch>=0.3.0",
]

docker = [
    "docker>=7.0.0",
]

all = [
    "frontend-tester[dev,ai,visual,docker]",
]

[project.scripts]
frontend-tester = "frontend_tester.cli.main:app"

[project.urls]
Homepage = "https://github.com/yourusername/frontend-tester"
Documentation = "https://github.com/yourusername/frontend-tester/docs"
Repository = "https://github.com/yourusername/frontend-tester"
Issues = "https://github.com/yourusername/frontend-tester/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
